CLI Utility to generate JSON object data (Initial)
Allows for number of fields to be specified, along with each fields name, whether or not there is a list of accepted values to put for that field (i.e in the case of this using ID's from other inserted MongoDB objects is there is a list of those, etc.).

If a list isn't provided it is then asked how many values for that field should be generated as well as if they should be HTML strings or not (if not it will simply generate a regular random string), if a specified accepted value list can't be found an empty array is generated.

If less values than the maximum values generated for any field for a particular field subsequent entries are filled with an empty string.

File is then outputted as output.json in which it will be an array of objects with the proper generated data for each Object field.